+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🐛 fix: children rules can now have no value (they had before for some reason)
 1.3.3 (2023-07-10)
 ------------------